modularization and plyer
-Funções foram separadas do main
-win10toast foi substituido pelo Plyer
