--- historico_precos/ANIMAL WELL_price_history.csv
@@ -1,5 +1,0 @@
-data,preco,jogo
-2025-02-12,"63,74",ANIMAL WELL
-2025-02-12,"63,74",ANIMAL WELL
-2025-02-12,"63,74",ANIMAL WELL
-2025-02-12,"63,74",ANIMAL WELL
