--- historico_precos/CHRONO CROSS THE RADICAL DREAMERS EDITION_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"52,93",CHRONO CROSS THE RADICAL DREAMERS EDITION
-2025-02-12,"52,93",CHRONO CROSS THE RADICAL DREAMERS EDITION
-2025-02-12,"52,93",CHRONO CROSS THE RADICAL DREAMERS EDITION
-2025-02-12,"52,93",CHRONO CROSS THE RADICAL DREAMERS EDITION
-2025-02-12,"52,93",CHRONO CROSS THE RADICAL DREAMERS EDITION
