--- historico_precos/Cult of the Lamb_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"36,99",Cult of the Lamb
-2025-02-12,"36,99",Cult of the Lamb
-2025-02-12,"36,99",Cult of the Lamb
-2025-02-12,"36,99",Cult of the Lamb
-2025-02-12,"36,99",Cult of the Lamb
