--- historico_precos/DAVE THE DIVER_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"63,15",DAVE THE DIVER
-2025-02-12,"63,15",DAVE THE DIVER
-2025-02-12,"63,15",DAVE THE DIVER
-2025-02-12,"63,15",DAVE THE DIVER
-2025-02-12,"63,15",DAVE THE DIVER
