--- historico_precos/Pepper Grinder_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"28,19",Pepper Grinder
-2025-02-12,"28,19",Pepper Grinder
-2025-02-12,"28,19",Pepper Grinder
-2025-02-12,"28,19",Pepper Grinder
-2025-02-12,"28,19",Pepper Grinder
