--- historico_precos/Prince of Persia The Lost Crown_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"104,51",Prince of Persia The Lost Crown
-2025-02-12,"104,51",Prince of Persia The Lost Crown
-2025-02-12,"104,51",Prince of Persia The Lost Crown
-2025-02-12,"104,51",Prince of Persia The Lost Crown
-2025-02-12,"104,51",Prince of Persia The Lost Crown
