--- historico_precos/Red Dead Redemption_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"153,83",Red Dead Redemption
-2025-02-12,"153,83",Red Dead Redemption
-2025-02-12,"153,83",Red Dead Redemption
-2025-02-12,"153,83",Red Dead Redemption
-2025-02-12,"153,83",Red Dead Redemption
