--- historico_precos/SONIC X SHADOW GENERATIONS_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"204,01",SONIC X SHADOW GENERATIONS
-2025-02-12,"204,01",SONIC X SHADOW GENERATIONS
-2025-02-12,"204,01",SONIC X SHADOW GENERATIONS
-2025-02-12,"173,37",SONIC X SHADOW GENERATIONS
-2025-02-12,"173,37",SONIC X SHADOW GENERATIONS
