--- historico_precos/Tales of Symphonia Remastered_price_history.csv
@@ -1,6 +1,0 @@
-data,preco,jogo
-2025-02-12,"71,89",Tales of Symphonia Remastered
-2025-02-12,"71,89",Tales of Symphonia Remastered
-2025-02-12,"71,89",Tales of Symphonia Remastered
-2025-02-12,"71,89",Tales of Symphonia Remastered
-2025-02-12,"71,89",Tales of Symphonia Remastered
